+++ src/lib/replacement/javascript.js
@@ -4,8 +4,10 @@ function replace(fileContent, replacementMap) {
   return eol
     .split(fileContent)
     .map(line => treatRequire(line, replacementMap))
+    .map(line => treatRequireActual(line, replacementMap))
     .map(line => treatImport(line, replacementMap))
     .map(line => treatNodeModules(line, replacementMap))
+    .map(line => treatJest(line, replacementMap))
     .join(eol.auto);
 }
 
@@ -17,16 +19,32 @@ function treatRequire(line, replacementMap) {
   );
 }
 
+function treatRequireActual(line, replacementMap) {
+  return treatRegEx(
+    line,
+    replacementMap,
+    `(require\\.requireActual\\(["|'])($KEY)(.*["|']\\))(;)?`
+  );
+}
+
 function treatImport(line, replacementMap) {
   return treatRegEx(
     line,
     replacementMap,
-    `(import { .* } from ["|'])($KEY)(.*["|'])(;)?`
+    `((})? from ["|'])($KEY)(.*["|'])(;)?`
   );
 }
 
 function treatNodeModules(line, replacementMap) {
   return treatRegEx(line, replacementMap, `(node_modules/)($KEY)(.*)(;)?`);
+}
+
+function treatJest(line, replacementMap) {
+  return treatRegEx(
+    line,
+    replacementMap,
+    `(jest.mock\\(["|'])($KEY)(.*["|'])(;)?`
+  );
 }
 
 function treatRegEx(line, replacementMap, regEx) {