+++ src/lib/replacement/package.json.js
@@ -22,9 +22,10 @@ function treatScripts(scripts, replacementMap) {
   if (scripts && replacementMap) {
     Object.entries(scripts).forEach(([scriptKey, scriptValue]) =>
       Object.entries(replacementMap).forEach(([key, value]) => {
-        let script = scriptValue.replace(new RegExp(`\s${key}`, "gi"), value);
+        let script = scriptValue.replace(new RegExp(` ${key}`, "gi"), value);
         script = script.replace(new RegExp(`=${key}`, "gi"), `=${value}`);
         scripts[scriptKey] = script;
+        scriptValue = script;
       })
     );
   }