+++ test/resources/testfile2.js
@@ -76,11 +76,7 @@ module.exports = async (env, argv) => {
         { from: "./static/samples", to: "./samples" },
         { from: "./static/index.html", to: "./index.html" },
         { from: "./static/favicon.ico", to: "./favicon.ico" },
-        {
-          from:
-            "../../node_modules/@kogito-tooling/kie-bc-editors/dist/envelope-dist",
-          to: "./envelope"
-        },
+        { from: "../../node_modules/@kogito-tooling/kie-bc-editors/dist/envelope-dist", to: "./envelope" },
         { from: externalAssets.dmnEditorPath(argv), to: "./gwt-editors/dmn" },
         { from: externalAssets.bpmnEditorPath(argv), to: "./gwt-editors/bpmn" }
       ])