+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 # V1.0
 
-- add/rewrite scope from package.json name
+- add/rewrite scope from package.json, .ts, .tsx and .js files.